+++ models/base_trainer.py
@@ -5,10 +5,13 @@ import torch.nn as nn
 class BaseTrainer(nn.Module):
     def __init__(self):
         super().__init__()
+
     def build_model(self):
         raise NotImplementedError
+
     def train_on_batch(self, X, optimizer, loss_manager, valid_mode=False, device=torch.device('cpu')):
         raise NotImplementedError
+
     def update_nn_weights(self, optimizer, loss_manager, valid_mode=False):
         loss_manager.updateLossHistory()
         loss = loss_manager.computeTotalLoss()
