+++ models/gan.py
@@ -2,236 +2,18 @@ import numpy as np
 import torch
 from torch import nn
 import torch.nn.functional as F
-from torch.nn.utils import spectral_norm
 try:
     # relative import
-    from .models import BaseModelSRL
+    from .base_models import BaseModelSRL, ConvSN2d, ConvTransposeSN2d, LinearSN, UNet
     from .base_trainer import BaseTrainer
     from ..losses.losses import ganNonSaturateLoss, autoEncoderLoss, ganBCEaccuracy, AEboundLoss
 except:
-    from models.models import BaseModelSRL
+    from models.base_models import BaseModelSRL, ConvSN2d, ConvTransposeSN2d, LinearSN, UNet
     from models.base_trainer import BaseTrainer
     from losses.losses import ganNonSaturateLoss, autoEncoderLoss, ganBCEaccuracy, AEboundLoss
     
-def ConvSN2d(in_channels, out_channels, kernel_size,
-             stride=1,
-             padding=0,
-             dilation=1,
-             groups=1,
-             bias=True,
-             padding_mode='zeros'):
-    A = spectral_norm(nn.Conv2d(in_channels, out_channels,
-                                kernel_size=kernel_size,
-                                stride=stride,
-                                padding=padding,
-                                dilation=dilation,
-                                groups=groups,
-                                bias=bias))
-    # A.__class__.__name__ = 'ConvSN2d' ## [TODO]
-    return A
-
-
-def ConvTransposeSN2d(in_channels, out_channels, kernel_size,
-                      stride=1,
-                      padding=0,
-                      output_padding=0,
-                      groups=1,
-                      bias=True,
-                      dilation=1,
-                      padding_mode='zeros'):
-    A = spectral_norm(nn.ConvTranspose2d(in_channels, out_channels,
-                                         kernel_size=kernel_size,
-                                         stride=stride,
-                                         padding=padding,
-                                         output_padding=output_padding,
-                                         groups=groups,
-                                         bias=bias,
-                                         dilation=dilation))
-    # A.__class__.__name__ = 'ConvTransposeSN2d' ## [TODO]
-    return A
-
-
-def LinearSN(in_features, out_features, bias=True):
-    A = spectral_norm(nn.Linear(in_features, out_features, bias=bias))
-    # A.__class__.__name__ = 'LinearSN' ## [TODO]
-    return A
-
-
-class UNet(nn.Module):
-    def __init__(
-        self,
-        in_ch=3,
-        out_ch=3,
-        depth=4,
-        start_ch=64,
-        inc_rate=2,
-        padding=True,
-        batch_norm=True,
-        spec_norm=False,
-        dropout=0.5,
-        up_mode='upconv',
-        include_top=True
-    ):
-        """
-        Implementation of
-        U-Net: Convolutional Networks for Biomedical Image Segmentation
-        (Ronneberger et al., 2015)
-        https://arxiv.org/abs/1505.04597
-        Args:
-            in_ch (int): number of input channels
-            out_ch (int): number of output channels
-            depth (int): depth of the network
-            start_ch (int): number of filters in the first layer is start_ch
-            padding (bool): if True, apply padding such that the input shape
-                            is the same as the output.
-                            This may introduce artifacts
-            batch_norm (bool): Use BatchNorm after layers with an
-                               activation function
-            dropout (None or float): Use dropout (if not None) in Conv block.
-            up_mode (str): one of 'upconv' or 'upsample'.
-                           'upconv' will use transposed convolutions for
-                           learned upsampling.
-                           'upsample' will use bilinear upsampling.
-        """
-        super(UNet, self).__init__()
-        assert up_mode in ('upconv', 'upsample')
-        self.include_top = include_top
-        self.padding = padding
-        self.out_ch = out_ch
-        self.depth = depth
-        self.spec_norm = spec_norm
-        prev_channels = in_ch
-        self.down_path = nn.ModuleList()
-        for i in range(depth+1):
-            self.down_path.append(
-                UNetConvBlock(prev_channels, (inc_rate ** i) * start_ch,
-                              padding, batch_norm, dropout, spec_norm=self.spec_norm)
-            )
-            prev_channels = (inc_rate ** i) * start_ch
-
-        self.up_path = nn.ModuleList()
-        for i in reversed(range(depth)):
-            self.up_path.append(
-                UNetUpBlock(prev_channels, (inc_rate ** i) * start_ch, up_mode,
-                            padding, batch_norm, dropout, spec_norm=self.spec_norm)
-            )
-            prev_channels = (inc_rate ** i) * start_ch
-
-        if self.include_top:
-            if self.spec_norm:
-                self.last = ConvSN2d(prev_channels, out_ch, kernel_size=1)
-            else:
-                self.last = nn.Conv2d(prev_channels, out_ch, kernel_size=1)
-        else:
-            self.out_ch = prev_channels
-        self.tanh_act = nn.Tanh()
-
-    def forward(self, x):
-        blocks = []
-        for i, down in enumerate(self.down_path):
-            x = down(x)
-            if i != len(self.down_path) - 1:
-                blocks.append(x)
-                x = F.max_pool2d(x, 2)
-
-        for i, up in enumerate(self.up_path):
-            x = up(x, blocks[-i - 1])
-
-        if self.include_top:
-            x = self.last(x)
-            return self.tanh_act(x)
-        else:
-            return x
-
-
-class UNetConvBlock(nn.Module):
-    def __init__(self, in_size, out_size, padding, batch_norm, dropout, spec_norm=False):
-        super(UNetConvBlock, self).__init__()
-        self.spec_norm = spec_norm
-        block = []
-        if self.spec_norm:
-            # [stride=1] padding = (k-1)/2
-            block.append(ConvSN2d(in_size, out_size,
-                                  kernel_size=3, padding=int(padding)))
-        else:
-            # [stride=1] padding = (k-1)/2
-            block.append(nn.Conv2d(in_size, out_size,
-                                   kernel_size=3, padding=int(padding)))
-        block.append(nn.ReLU())
-        if batch_norm:
-            block.append(nn.BatchNorm2d(out_size))
-        if dropout is not None:
-            block.append(nn.Dropout(p=dropout))
-        if self.spec_norm:
-            block.append(ConvSN2d(out_size, out_size,
-                                  kernel_size=3, padding=int(padding)))
-        else:
-            block.append(nn.Conv2d(out_size, out_size,
-                                   kernel_size=3, padding=int(padding)))
-        block.append(nn.ReLU())
-        if batch_norm:
-            block.append(nn.BatchNorm2d(out_size))
-
-        self.block = nn.Sequential(*block)
-
-    def forward(self, x):
-        out = self.block(x)
-        return out
-
-
-class UNetUpBlock(nn.Module):
-    def __init__(self, in_size, out_size, up_mode, padding, batch_norm, dropout, spec_norm=False):
-        super(UNetUpBlock, self).__init__()
-        self.padding = padding
-        self.spec_norm = spec_norm
-        if up_mode == 'upconv':
-            if self.spec_norm:
-                self.up = ConvTransposeSN2d(
-                    in_size, out_size, kernel_size=3, stride=2, padding=1, output_padding=1)
-            else:
-                # self.up = nn.ConvTranspose2d(in_size, out_size, kernel_size=2, stride=2)
-                self.up = nn.ConvTranspose2d(
-                    in_size, out_size, kernel_size=3, stride=2, padding=1, output_padding=1)
-        elif up_mode == 'upsample':
-
-            if self.spec_norm:
-                self.up = nn.Sequential(
-                    nn.Upsample(mode='bilinear', scale_factor=2),
-                    ConvSN2d(in_size, out_size, kernel_size=1),
-                )
-            else:
-                self.up = nn.Sequential(
-                    nn.Upsample(mode='bilinear', scale_factor=2),
-                    nn.Conv2d(in_size, out_size, kernel_size=1),
-                )
-        self.relu_act = nn.ReLU()
-        self.conv_block = UNetConvBlock(
-            in_size, out_size, padding, batch_norm, dropout, spec_norm=self.spec_norm)
-
-    def center_crop(self, layer, target_size):
-        _, _, layer_height, layer_width = layer.size()
-        diff_y = (layer_height - target_size[0]) // 2
-        diff_x = (layer_width - target_size[1]) // 2
-        return layer[:, :, diff_y: (diff_y + target_size[0]), diff_x: (diff_x + target_size[1])]
-
-    def forward(self, x, bridge):
-        up = self.up(x)
-        up = self.relu_act(up)
-        # print("Brige shape: {}, target size: {}".format(bridge.shape, up.shape[2:]))
-        if self.padding:
-            crop1 = bridge
-        else:
-            crop1 = self.center_crop(bridge, up.shape[2:])
-        # print(up.shape)
-        # print(crop1.shape)
-        out = torch.cat([up, crop1], 1)
-        out = self.conv_block(out)
-
-        return out
-
-
 class Generator(nn.Module):
-    def __init__(self, img_shape, state_dim,
+    def __init__(self, state_dim, img_shape,
                  unet_depth=2, # 3
                  unet_ch=16, # 32
                  spectral_norm=False,
@@ -286,7 +68,7 @@ class Generator(nn.Module):
 
 
 class Discriminator(nn.Module):
-    def __init__(self, img_shape, state_dim,
+    def __init__(self, state_dim, img_shape,
                  spectral_norm=False,
                  d_chs=16): # 32
         super().__init__()
@@ -369,7 +151,8 @@ class Encoder(BaseModelSRL):
     
     Note: Only Encoder has getStates method.
     """
-    def __init__(self, img_shape, state_dim,
+
+    def __init__(self, state_dim, img_shape,
                  unet_depth=2, # 3 
                  unet_ch=16,
                  unet_bn=False,
@@ -442,9 +225,9 @@ class GANTrainer(BaseTrainer):
         self.img_shape = img_shape
         self.state_dim = state_dim
     def build_model(self):
-        self.encoder = Encoder(self.img_shape, self.state_dim, spectral_norm=False)
-        self.generator = Generator(self.img_shape, self.state_dim, spectral_norm=True)
-        self.discriminator = Discriminator(self.img_shape, self.state_dim, spectral_norm=True)
+        self.encoder = Encoder(self.state_dim, self.img_shape, spectral_norm=False)
+        self.generator = Generator(self.state_dim, self.img_shape, spectral_norm=True)
+        self.discriminator = Discriminator(self.state_dim, self.img_shape, spectral_norm=True)
 
     def forward(self, x):
         return self.encoder(x)
@@ -507,11 +290,11 @@ if __name__ == "__main__":
     # a = 128
     # summary(model, (chs, a, a))
 
-    # model = Generator(img_shape=(3, 128, 128), 10)
+    # model = Generator(10, img_shape=(3, 128, 128))
     # summary(model, (10,))
 
-    # model = Discriminator((3, 128, 128), 4)
+    # model = Discriminator(4, (3, 128, 128))
     # summary(model, (3, 128, 128))
 
-    model = Encoder((3,128,128), 4)
+    model = Encoder(4, (3, 128, 128))
     summary(model, (3, 128, 128))
