+++ models/triplet.py
@@ -1,6 +1,6 @@
 from __future__ import print_function, division, absolute_import
 
-from .models import *
+from .base_models import *
 
 
 class EmbeddingNet(BaseModelSRL):
